requirements.txt
astroid==2.2.5
bootstrap4==0.1.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.0.0
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.8.0
gunicorn==19.9.0
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.0.0
psycopg2==2.8.2
pylint==2.3.1
python-decouple==3.1
pytz==2019.1
six==1.12.0
typed-ast==1.3.5
whitenoise==4.1.2
wrapt==1.11.1
